+++ iOS~/eppz! Cloud/eppz! Cloud/EPPZ_Cloud_C.m
@@ -11,38 +11,38 @@
 // OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-#import "EPPZ_Cloud_C++.h"
+#import "EPPZ_Cloud_C.h"
 #import "UnityString.m"
 
 
-extern void EPPZ_Cloud_InitializeWithGameObjectName(NSString* gameObjectName)
+extern void _EPPZ_Cloud_InitializeWithGameObjectName(NSString* gameObjectName)
 { [[EPPZ_Cloud instance] initializeWithGameObjectName:UnityStringFromNSString(gameObjectName)]; }
 
-extern bool EPPZ_Cloud_Synchronize(void)
+extern bool _EPPZ_Cloud_Synchronize(void)
 { return [[EPPZ_Cloud instance] synchronize]; }
 
-extern NSString* EPPZ_Cloud_StringForKey(NSString* key)
+extern NSString* _EPPZ_Cloud_StringForKey(NSString* key)
 { return NSStringFromUnityString([[EPPZ_Cloud instance] stringForKey:UnityStringFromNSString(key)]); }
 
-extern void EPPZ_Cloud_SetStringForKey(NSString* value, NSString* key)
+extern void _EPPZ_Cloud_SetStringForKey(NSString* value, NSString* key)
 { [[EPPZ_Cloud instance] setString:UnityStringFromNSString(value) forKey:UnityStringFromNSString(key)]; }
 
-extern float EPPZ_Cloud_FloatForKey(NSString* key)
+extern float _EPPZ_Cloud_FloatForKey(NSString* key)
 { return [[EPPZ_Cloud instance] floatForKey:UnityStringFromNSString(key)]; }
 
-extern void EPPZ_Cloud_SetFloatForKey(float value, NSString* key)
+extern void _EPPZ_Cloud_SetFloatForKey(float value, NSString* key)
 { [[EPPZ_Cloud instance] setFloat:value forKey:UnityStringFromNSString(key)]; }
 
-extern int EPPZ_Cloud_IntForKey(NSString* key)
+extern int _EPPZ_Cloud_IntForKey(NSString* key)
 { return [[EPPZ_Cloud instance] intForKey:UnityStringFromNSString(key)]; }
 
-extern void EPPZ_Cloud_SetIntForKey(int value, NSString* key)
+extern void _EPPZ_Cloud_SetIntForKey(int value, NSString* key)
 { [[EPPZ_Cloud instance] setInt:value forKey:UnityStringFromNSString(key)]; }
 
-extern bool EPPZ_Cloud_BoolForKey(NSString* key)
+extern bool _EPPZ_Cloud_BoolForKey(NSString* key)
 { return [[EPPZ_Cloud instance] boolForKey:UnityStringFromNSString(key)]; }
 
-extern void EPPZ_Cloud_SetBoolForKey(bool value, NSString* key)
+extern void _EPPZ_Cloud_SetBoolForKey(bool value, NSString* key)
 { [[EPPZ_Cloud instance] setBool:value forKey:UnityStringFromNSString(key)]; }